+++ packages/core/build.config.ts
@@ -4,6 +4,16 @@ import { defineBuildConfig } from 'unbuild'
 export default defineBuildConfig( [
 	{
 		...config,
-		externals : [ 'deepmerge-ts' ],
+		externals : [
+			'deepmerge-ts',
+			'cli-table3',
+			'gradient-string',
+			'string-width',
+			'tiny-glob',
+			'columnify',
+			'@clack/prompts',
+			'@clack/core',
+			'@visulima/boxen',
+		],
 	},
 ] )