+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # creatium
 
+## 0.2.1
+
+### Patch Changes
+
+- fix types, remove common support, fix issues
+
 ## 0.2.0
 
 ### Minor Changes