+++ packages/create/CHANGELOG.md
@@ -1,5 +1,14 @@
 # create-creatium
 
+## 0.2.1
+
+### Patch Changes
+
+- fix types, remove common support, fix issues
+
+- Updated dependencies []:
+  - creatium@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes